pyproject.toml
[project]
name = "synco"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.12,<6",
    "django-htmx>=1.27.0",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.1.0",
    "dj-database-url>=2.3.0",
    "httpx>=0.28.0",
    "django-widget-tweaks>=1.5.0",
    "gunicorn>=23.0.0",
    "whitenoise>=6.9.0",
    "openpyxl>=3.1.0",
    "google-genai>=1.0.0",
    "pgvector>=0.4.0",
    "openai>=2.30.0",
    "google-api-python-client>=2.193.0",
    "google-auth-oauthlib>=1.3.0",
    "python-docx>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11.0",
    "pytest>=8.0.0",
    "pytest-django>=4.9.0",
]

[tool.ruff]
exclude = ["docs/"]

[tool.ruff.lint]
ignore = ["E402"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "main.settings"
pythonpath = ["."]
